+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 
 > **LoginRadius Node JS SDK Change Log** provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention](https://docs.loginradius.com/api/v2/deployment/sdk-libraries/node-js-library)
 
+# Version 11.1.0
+Release on April 7, 2021
+
+## Enhancements
+
+ - Added X-Origin-IP header support.
+ - Added 429 error code handling for "Too Many Request in a particular time frame".
+
+
+## Added new multiple APIs for better user experience
+
+ - Get Profile By Ping.
+ - Passwordless Login Verification By Email And OTP.
+ - Passwordless Login Verification By User Name And OTP.
+
+
+
 # Version 11.0.0
 Release on Aug 04, 2020